perf(ui): defer native plugin contribution rendering
