+++ ui/config/control-ui-boot-modules.json
@@ -931,6 +931,7 @@
   "src/gateway/control-reply-text.ts",
   "src/gateway/control-ui-bootstrap-contract.ts",
   "src/gateway/control-ui-contract.ts",
+  "src/gateway/control-ui-plugin-assets-contract.ts",
   "src/gateway/control-ui-plugin-frame-contract.ts",
   "src/gateway/control-ui-resource-routes.ts",
   "src/gateway/control-ui-root-assets.ts",
@@ -946,9 +947,11 @@
   "src/media/media-facts.ts",
   "src/media/parse.ts",
   "src/plugin-sdk/message-tool-delivery-hints.ts",
+  "src/shared/agent-runtime-display.ts",
   "src/shared/avatar-limits.ts",
   "src/shared/chat-envelope.ts",
   "src/shared/chat-message-content.ts",
+  "src/shared/deferred.ts",
   "src/shared/device-auth.ts",
   "src/shared/device-bootstrap-profile.ts",
   "src/shared/fast-mode.ts",
@@ -959,6 +962,7 @@
   "src/shared/regexp.ts",
   "src/shared/session-archive-timeout.ts",
   "src/shared/session-method-scopes-base.ts",
+  "src/shared/session-run-state.ts",
   "src/shared/text/assistant-visible-text.ts",
   "src/shared/text/citation-control-markers.ts",
   "src/shared/text/code-regions.ts",
@@ -1300,7 +1304,6 @@
   "ui/src/lib/session-method-access.ts",
   "ui/src/lib/session-model-defaults.ts",
   "ui/src/lib/session-pull-requests.ts",
-  "ui/src/lib/session-run-state.ts",
   "ui/src/lib/sessions/catalog-key.ts",
   "ui/src/lib/sessions/catalog-project-grouping.ts",
   "ui/src/lib/sessions/catalog-terminal.ts",
@@ -1480,6 +1483,7 @@
   "ui/src/pages/dashboards/route.ts",
   "ui/src/pages/debug/debug-overlay-contract.ts",
   "ui/src/pages/debug/route.ts",
+  "ui/src/pages/device/route.ts",
   "ui/src/pages/devices/route.ts",
   "ui/src/pages/labs/route.ts",
   "ui/src/pages/lobsterdex/route.ts",
@@ -1635,9 +1639,9 @@
   "src/shared/agent-run-status.ts",
   "src/shared/bounded-buffer.ts",
   "src/shared/bounded-serial-queue.ts",
-  "src/shared/deferred.ts",
   "src/shared/markdown-code.ts",
   "src/shared/resume-handoff.ts",
+  "src/shared/session-goal-display.ts",
   "src/shared/transcript-only-openclaw-assistant.ts",
   "src/talk/agent-consult-tool.ts",
   "src/talk/agent-run-control-shared.ts",
@@ -1693,7 +1697,6 @@
   "ui/src/lib/chat/tool-display.ts",
   "ui/src/lib/provider-quota-summary.ts",
   "ui/src/lib/response-body.ts",
-  "ui/src/lib/session-goal.ts",
   "ui/src/lib/session-progress-cards.ts",
   "ui/src/lib/session-viewer-presence.ts",
   "ui/src/lib/sessions/session-archive-reason.ts",
@@ -1758,6 +1761,7 @@
   "ui/src/pages/chat/chat-pull-request-refresh.ts",
   "ui/src/pages/chat/chat-realtime.ts",
   "ui/src/pages/chat/chat-reset-delivery.ts",
+  "ui/src/pages/chat/chat-scroll-input.ts",
   "ui/src/pages/chat/chat-send-actions.ts",
   "ui/src/pages/chat/chat-send-composer.ts",
   "ui/src/pages/chat/chat-send-delivery.ts",
